GemmAI/requirements.txt
Flask
google.genai
dotenv
Flask-WTF
Flask-SqlAlchemy
Flask-migrate
Flask-login
gunicorn
